# Conflicts:
#	.idea/modules.xml

+++ app/src/main/java/com/android/policyboss/core/models/OwnDamageEntity.java
@@ -1,29 +1,29 @@
 package com.android.policyboss.core.models;
 
-public  class OwnDamageEntity {
+public class OwnDamageEntity {
                 /**
-                 * od_basic : 1127.16
+                 * od_basic : 11490.07
                  * od_elect_access : 0
                  * od_non_elect_access : 0
                  * od_cng_lpg : 0
-                 * od_disc_ncb : 211.34
+                 * od_disc_ncb : 2298.01
                  * od_disc_vol_deduct : 0
                  * od_disc_anti_theft : 0
                  * od_disc_aai : 0
                  * od_loading : 0
-                 * od_disc : 281.79
-                 * od_final_premium : 634.03
+                 * od_disc : 0
+                 * od_final_premium : 9192.06
                  */
 
                 private double od_basic;
-                private int od_elect_access;
-                private int od_non_elect_access;
-                private int od_cng_lpg;
+                private double od_elect_access;
+                private double od_non_elect_access;
+                private double od_cng_lpg;
                 private double od_disc_ncb;
-                private int od_disc_vol_deduct;
-                private int od_disc_anti_theft;
-                private int od_disc_aai;
-                private int od_loading;
+                private double od_disc_vol_deduct;
+                private double od_disc_anti_theft;
+                private double od_disc_aai;
+                private double od_loading;
                 private double od_disc;
                 private double od_final_premium;
 
@@ -35,27 +35,27 @@ public  class OwnDamageEntity {
                     this.od_basic = od_basic;
                 }
 
-                public int getOd_elect_access() {
+                public double getOd_elect_access() {
                     return od_elect_access;
                 }
 
-                public void setOd_elect_access(int od_elect_access) {
+                public void setOd_elect_access(double od_elect_access) {
                     this.od_elect_access = od_elect_access;
                 }
 
-                public int getOd_non_elect_access() {
+                public double getOd_non_elect_access() {
                     return od_non_elect_access;
                 }
 
-                public void setOd_non_elect_access(int od_non_elect_access) {
+                public void setOd_non_elect_access(double od_non_elect_access) {
                     this.od_non_elect_access = od_non_elect_access;
                 }
 
-                public int getOd_cng_lpg() {
+                public double getOd_cng_lpg() {
                     return od_cng_lpg;
                 }
 
-                public void setOd_cng_lpg(int od_cng_lpg) {
+                public void setOd_cng_lpg(double od_cng_lpg) {
                     this.od_cng_lpg = od_cng_lpg;
                 }
 
@@ -67,35 +67,35 @@ public  class OwnDamageEntity {
                     this.od_disc_ncb = od_disc_ncb;
                 }
 
-                public int getOd_disc_vol_deduct() {
+                public double getOd_disc_vol_deduct() {
                     return od_disc_vol_deduct;
                 }
 
-                public void setOd_disc_vol_deduct(int od_disc_vol_deduct) {
+                public void setOd_disc_vol_deduct(double od_disc_vol_deduct) {
                     this.od_disc_vol_deduct = od_disc_vol_deduct;
                 }
 
-                public int getOd_disc_anti_theft() {
+                public double getOd_disc_anti_theft() {
                     return od_disc_anti_theft;
                 }
 
-                public void setOd_disc_anti_theft(int od_disc_anti_theft) {
+                public void setOd_disc_anti_theft(double od_disc_anti_theft) {
                     this.od_disc_anti_theft = od_disc_anti_theft;
                 }
 
-                public int getOd_disc_aai() {
+                public double getOd_disc_aai() {
                     return od_disc_aai;
                 }
 
-                public void setOd_disc_aai(int od_disc_aai) {
+                public void setOd_disc_aai(double od_disc_aai) {
                     this.od_disc_aai = od_disc_aai;
                 }
 
-                public int getOd_loading() {
+                public double getOd_loading() {
                     return od_loading;
                 }
 
-                public void setOd_loading(int od_loading) {
+                public void setOd_loading(double od_loading) {
                     this.od_loading = od_loading;
                 }
